MAINT: mass rename
