+++ euphoria-service/src/main/java/club/euphoria_recruiting/handler/ApplicationHandlers.java
@@ -5,8 +5,8 @@ import com.google.common.annotations.VisibleForTesting;
 import com.spotify.apollo.RequestContext;
 import com.spotify.apollo.Response;
 import com.spotify.apollo.route.*;
-import edu.cooper.ece366.euphoria.model.Application;
-import edu.cooper.ece366.euphoria.store.model.ApplicationStore;
+import club.euphoria_recruiting.model.Application;
+import club.euphoria_recruiting.store.model.ApplicationStore;
 import okio.ByteString;
 
 import java.io.IOException;
