+++ euphoria-service/src/main/java/club/euphoria_recruiting/model/Posting.java
@@ -1,8 +1,8 @@
-package edu.cooper.ece366.euphoria.model;
+package club.euphoria_recruiting.model;
 
-import edu.cooper.ece366.euphoria.utils.Industry;
-import edu.cooper.ece366.euphoria.utils.Location;
-import edu.cooper.ece366.euphoria.utils.SkillLevel;
+import club.euphoria_recruiting.utils.Industry;
+import club.euphoria_recruiting.utils.Location;
+import club.euphoria_recruiting.utils.SkillLevel;
 import io.norberg.automatter.AutoMatter;
 
 @AutoMatter
