+++ euphoria-service/src/main/java/club/euphoria_recruiting/utils/EducationLevel.java
@@ -1,4 +1,4 @@
-package edu.cooper.ece366.euphoria.utils;
+package club.euphoria_recruiting.utils;
 
 public enum EducationLevel {
     NOHIGHSCHOOL, HIGHSCHOOL, GED, SOMECOLLEGE, ASSOCIATES, BACHELORS, MASTERS, PHD, MD, JD
