+++ euphoria-service/src/main/java/club/euphoria_recruiting/utils/Industry.java
@@ -1,4 +1,4 @@
-package edu.cooper.ece366.euphoria.utils;
+package club.euphoria_recruiting.utils;
 
 public enum Industry {
     EDUCATION, ENERGY, FINANCE, FOOD, HEALTHCARE, INSURANCE, MEDIA, RETAIL, SERVICES, TECHNOLOGY, TRANSPORT, UTILITIES
