+++ euphoria-service/src/main/java/club/euphoria_recruiting/utils/Location.java
@@ -1,4 +1,4 @@
-package edu.cooper.ece366.euphoria.utils;
+package club.euphoria_recruiting.utils;
 
 public enum Location {
     NEWYORK, LONDON, HONGKONG, BERLIN, BEIJING, WASHINGTON
