+++ euphoria-service/src/test/java/club/euphoria_recruiting/PostingHandlersTest.java
@@ -1,15 +1,15 @@
-package edu.cooper.ece366.euphoria;
+package club.euphoria_recruiting;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.spotify.apollo.Request;
 import com.spotify.apollo.RequestContext;
-import edu.cooper.ece366.euphoria.handler.PostingHandlers;
-import edu.cooper.ece366.euphoria.model.Posting;
-import edu.cooper.ece366.euphoria.model.PostingBuilder;
-import edu.cooper.ece366.euphoria.store.model.PostingStore;
-import edu.cooper.ece366.euphoria.utils.Industry;
-import edu.cooper.ece366.euphoria.utils.Location;
-import edu.cooper.ece366.euphoria.utils.SkillLevel;
+import club.euphoria_recruiting.handler.PostingHandlers;
+import club.euphoria_recruiting.model.Posting;
+import club.euphoria_recruiting.model.PostingBuilder;
+import club.euphoria_recruiting.store.model.PostingStore;
+import club.euphoria_recruiting.utils.Industry;
+import club.euphoria_recruiting.utils.Location;
+import club.euphoria_recruiting.utils.SkillLevel;
 import okio.ByteString;
 import org.junit.Before;
 import org.junit.Test;
